--- src/geral.css
@@ -1,3 +1,0 @@
-body {
-  color: brown;
-}
