+++ src/variables.css
@@ -1,0 +1,63 @@
+@import url('https://fonts.googleapis.com/css2?family=Roboto:wght@400;700&display=swap');
+
+/* AINDA NÃO USADO */
+
+:root {
+  /* Fonts */
+  /* Heading */
+  --font-heading-1: 36px;
+  --font-heading-2: 24px;
+  --font-heading-3: 18px;
+  
+  /* Body */
+  --font-body: 16px;
+  --font-body-bold: 16px;
+  --font-body-s: 14px;
+  --font-body-s-bold: 14px;
+  --font-body-l: 18px;
+  --font-body-l-bold: 18px;
+  
+  /* Button */
+  --font-button: 16px;
+  --font-button-s: 14px;
+  
+  /* Font Weights */
+  --font-weight-regular: 400;
+  --font-weight-bold: 700;
+  
+  /* Spacing */
+  --spacing-4: 4px;
+  --spacing-8: 8px;
+  --spacing-12: 12px;
+  --spacing-16: 16px;
+  --spacing-20: 20px;
+  --spacing-24: 24px;
+  --spacing-28: 28px;
+  --spacing-32: 32px;
+  --spacing-36: 36px;
+  --spacing-40: 40px;
+  --spacing-44: 44px;
+  --spacing-48: 48px;
+  --spacing-52: 52px;
+  --spacing-56: 56px;
+  --spacing-60: 60px;
+  
+  /* Button Colors */
+  --button-color-normal: #007BFF;
+  --button-color-hover: #0056D2;
+  --button-color-active: #003D99;
+  --button-color-disabled: #A3C3FF;
+  --button-color-text: #fff;
+  
+  /* Button Padding */
+  --button-padding-width: 14px;
+  --button-padding-height: 10px;
+
+  /* Font family */
+  --font-family: 'Roboto', sans-serif;
+}
+
+/* Global */
+body {
+  font-family: var(--font-family);
+}
